🔨 refactor: move todo-item in the root directory of todos for much cleaner structure
