+++ src/features/todos/index.ts
@@ -1,3 +1,4 @@
 export * from './add-todo-form'
 export * from './todo-context'
+export * from './todo-item'
 export * from './todo-list'
