+++ src/features/todos/todo-item/types.ts
@@ -1,4 +1,4 @@
-import type { Todo } from '../../types'
+import type { Todo } from '../types'
 
 export type TodoItemProps = OmitNonHTMLAttributes<JSX.IntrinsicElements['li']> & {
   todo: Todo
