+++ src/features/todos/types.ts
@@ -1,5 +1,6 @@
 export type * from './add-todo-form/types'
 export type * from './todo-context/types'
+export type * from './todo-item/types'
 export type * from './todo-list/types'
 
 export type Todo = {
